CMSIS-NN: Fix warnings
1. Doxygen warnings are fixed

2. Sign conversion warnings(-Wsign-conversion) and -Wextra warnings
   are fixed for TFLu compliant MVE APIs

+++ CMSIS/NN/Source/ActivationFunctions/arm_relu_q7.c
@@ -21,8 +21,8 @@
  * Title:        arm_relu_q7.c
  * Description:  Q7 version of ReLU
  *
- * $Date:        February 27, 2020
- * $Revision:    V.1.0.1
+ * $Date:        May 29, 2020
+ * $Revision:    V.1.0.2
  *
  * Target Processor:  Cortex-M cores
  *
@@ -69,7 +69,7 @@ void arm_relu_q7(q7_t *data, uint16_t size)
         in = read_q7x4_ia(&input);
 
         /* extract the first bit */
-        buf = __ROR(in & 0x80808080, 7);
+        buf = (int32_t)__ROR((uint32_t)in & 0x80808080, 7);
 
         /* if MSB=1, mask will be 0xFF, 0x0 otherwise */
         mask = __QSUB8(0x00000000, buf);
